export default daemonmanager
